Add boundary padding
Surround the map with padding on all sides, so it’s in the centre
of a 3x3 array of tiles when the diffusion computation is done.

I wonder if this is actually an improvement. Fixing the boundary
and using the Neumann boundary conditions that prevent flow through
the boundaries (as the cart software does) is an appealing approach
that produces good-looking results.