A new path exporter
The major innovation here is to treat shared borders as indivisible
units, thereby avoiding the skinny white trianges that can appear at
borders after path simplification. The natural next step is to vary
the degree of simplification based on the degree of maximal expansion
under any of the supplied cartogram grids.